Fix: Int8 overflow in time_minutes, Polars indexing in strategies
- data_loader.py: Cast dt.hour()/dt.minute() to Int32 before multiplication.
  Polars dt.hour() returns Int8 (max 127); 9*60=540 overflowed silently.
- expiry_day_pattern_v1.py: Use .to_list()[int(idx)] instead of Series[np.int64].
- vol_surface_signal_v1.py: Same fix + sample every 12th bar for efficiency.

Phase 3 audit results (5/5 strategies pass):
- Integration test: PASSED (first bar time_minutes=555, IST hour=9)
- implied_vol_dislocation_v1: 5601 CE, 340 PE signals
- implied_vs_realized_vol_v1: 5044 CE, 5045 PE signals
- intraday_vol_pattern_v1: 13 CE, 610 PE signals
- nifty_banknifty_spread_v1: 210 CE, 200 PE signals
- expiry_day_pattern_v1: 1291 CE, 1 PE signals

+++ pipeline/data_loader.py
@@ -76,9 +76,11 @@ def _add_time_columns(df: pl.DataFrame) -> pl.DataFrame:
         df = df.drop("_date")
 
     # time_minutes = minutes from midnight IST
+    # CRITICAL: dt.hour() returns Int8 (max 127). 9*60=540 overflows Int8.
+    # Must cast to Int32 BEFORE multiplication.
     df = df.with_columns(
-        (pl.col("datetime").dt.hour() * 60 + pl.col("datetime").dt.minute())
-        .cast(pl.Int32)
+        (pl.col("datetime").dt.hour().cast(pl.Int32) * 60
+         + pl.col("datetime").dt.minute().cast(pl.Int32))
         .alias("time_minutes")
     )
     return df
